Teach controller how to create and manage resources (#10)
When a new ClusterOrder is created, update .status.clusterReference and
create dependent resources (namespace, service account, rolebinding). When
a ClusterOrder is deleted, delete the namespace (and transitively delete
all the resources contained in the namespace).

Because we're creating resources in a namespace different from the one in
which the ClusterOrder is created, we're relying on labels to manage watch
events and trigger reconciliation.