Only respond to ClusterOrders in a specific namespace
We want to only watch ClusterOrder in a specific namespace. This will
ultimately simplify some aspects of the design and will effectively enforce
unique names on ClusterOrders (which is not the case when they can be
create in arbitrary namespaces).

The solution here is sub-optimal and will be replaced in a subsequence
change.