Update deployment manifests (#19)
* Add environment variables to manifests

Update the deployment manifest to use the namespace in which the operator
is deployed as CLOUDKIT_CLUSTER_ORDER_NAMESPACE, and optionall load webhook
environment variables from a secret.

We also set `imagePullPolicy: Always` to make life easier during
development.

* Bind manager service account to admin clusterrole

In order to grant the `admin` role to the per-cluster service accounts we
create, the manager service account must have that role cluster-wide.
Otherwise, creating the new rolebinding will fail with:

    rolebindings.rbac.authorization.k8s.io "cloudkit" is forbidden: user
    "system:serviceaccount:cloudkit-operator-system:cloudkit-operator-controller-manager"
    (groups=["system:serviceaccounts"
    "system:serviceaccounts:cloudkit-operator-system"
    "system:authenticated"]) is attempting to grant RBAC permissions not
    currently held: ...

* Update go version used to build container image