Add gRPC infrastructure (#23)
This patch adds a `buf.gen.yaml` file to generate the gRPC client code,
as well as the result of running `buf generate`.

This client code will be used in other patches to interact with the
fulfillment service API.

The patch also adds new `--grpc-plaintext`, `--grpc-insecure`,
`--grpc-token-file` and `--fulfillment-server-address` command line options to
specify the location of the fulfillment API server, and small piece of
code that lists the availble templates during startup, just as an
example to verify that the connection is working:

```
$ bin/manager --grpc-insecure --grpc-token-file token.txt --fulfillment-server-address fulfillment-api-innabox.apps...:443
2025-04-07T17:40:05+02:00       INFO    setup   gRPC connection is enabled
2025-04-07T17:40:06+02:00       INFO    setup   Available template      {"id": "ocp_4_17_small", "title": "OpenShift 4.17 small", "description": "OpenShift 4.17 with `small` instances as worker nodes."}
```

This will be disabled if the `--fulfillment-server-address` flag isn't used.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>