Send updates to the fulfillment service (#27)
* Send updates to the fulfillment service

This patch adds a new _Feedback_ controller that watches the Kubernetes
`ClusterOrder` object and sends sends updates to the fulfillment
service.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* Fix a bunch of linting issues

This mostly modifies variable names to make the linters happy.

I have disabled the `nakedret` linter, which I think is unnecessary, and
the `unparam` linter, which generates false positives.

---------

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Co-authored-by: Lars Kellogg-Stedman <lars@redhat.com>