Only list resources in a single namespace
Previously, we were calling `client.List(...)` without a namespace
restriction, which meant it was searching for matching resources across all
available namespaces. The associated label selector means that we weren't
going to select anything erroneously, but we were potentially increasing
the work required.

This commit limits all list operations to a single namespace.